refactor(mllp-client): flatten runtime/node.ts → node.ts
The package is now 7 flat modules — client, message, state, errors, queue
(unwired), index, node. Updated the package.json `./node` export
(dist/runtime/node.js → dist/node.js), the tsdown entry, and the test import.

Full monorepo green (85/85 turbo tasks); @glion/cli resolves the /node subpath.
